remove gcc 4.1 compiler warnings on linux at daimi
Ignore-this: 344c46c0da6f14dac7d2d698d7217996

darcs-hash:20100803151846-3a0eb-255b554b8503f4fad4901c380a0b046926eb084d.gz

+++ src/Resources/Texture2D.h
@@ -188,8 +188,8 @@ namespace OpenEngine {
                 if ( 0 > y || y > 1)
                     throw Core::Exception("y out of range");
 #endif
-                unsigned int X = x * width;
-                unsigned int Y = y * height;
+                unsigned int X = (unsigned int)(x * width);
+                unsigned int Y = (unsigned int)(y * height);
 
                 float dX = X / (float)width - x;
                 float dY = Y / (float)height - y;
